refactor(frontend): merge ScrollFade into ScrollFader
ScrollFade was only used internally by ScrollFader, so the
single-edge primitive had no external consumers earning the extra
file. Inline the scroll-edge detection and the two fade divs into
ScrollFader directly.
